Added more UI Controls and Implemented Functions
UI Controls were added and the functions for the radio buttons were implemented. I just need to set up a small window with a table of inactive companies to appear when ran and also possibly an add feature to add the desired company to track or add another feed for an existing company.

I'll get that done tomorrow and also look into unit testing methods.